+++ src/dashboard/utils.py
@@ -1,16 +1,9 @@
 """utils.py - Utility functions for the Streamlit app."""
 
-import sys
 import logging
 
-# Set up logging.
+# Get the logger instance.
 logger = logging.getLogger(__name__)
-logger.setLevel(logging.INFO)
-
-# Create a console handler and set the level to INFO.
-console_handler = logging.StreamHandler(sys.stdout)
-console_handler.setLevel(logging.INFO)
-logger.addHandler(console_handler)
 
 
 def is_streamlit_running() -> bool:
