+++ src/log_keeper/ingest.py
@@ -10,8 +10,7 @@ from pathlib import Path
 import pandas as pd
 from tqdm import tqdm
 
-# Set up logging.
-logging.basicConfig(level=logging.INFO)
+# Get the logger instance.
 logger = logging.getLogger(__name__)
